+++ docs/src/.vuepress/config.js
@@ -8,7 +8,7 @@ const docsSidebar = [
     title: "Concepts",
     collapsable: false,
     sidebarDepth: 2,
-    children: ["/docs/project", "/docs/plugins"],
+    children: ["/docs/singer-spec","/docs/project", "/docs/plugins"],
   },
   {
     title: "Guide",
@@ -164,7 +164,7 @@ module.exports = {
       },
       { text: "Blog", link: "https://www.meltano.com/blog", target: "_self" },
       {
-        text: "Join 900+ on Slack",
+        text: "Join 950+ on Slack",
         link:
           "https://join.slack.com/t/meltano/shared_invite/zt-obgpdeba-7yrqKhwyMBfdHDXsZY8G7Q",
         icon: "SlackIcon",